# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: Algorithmic Trading System API
  description: |
    Enterprise-grade algorithmic trading platform API providing comprehensive trading, portfolio management, 
    risk management, and market data services. Built with microservices architecture and event-driven design.
    
    ## Key Features
    - **Multi-Asset Trading**: Stocks, ETFs, Options, Futures, Forex, Crypto
    - **Real-time Market Data**: Live feeds with fallback mechanisms
    - **Advanced Risk Management**: Real-time monitoring, VaR, stress testing
    - **AI-Powered Analytics**: Strategy development and optimization
    - **High-Performance Execution**: Microsecond-level latency
    
    ## Authentication
    All API endpoints require authentication via JWT tokens or API keys.
    
    ## Rate Limiting
    - Standard endpoints: 1000 requests/minute
    - Market data endpoints: 10000 requests/minute
    - Trading endpoints: 500 requests/minute
    
    ## Error Handling
    The API uses standard HTTP status codes and returns detailed error information in JSON format.
  version: 1.0.0
  contact:
    name: Trading System API Support
    email: api-support@trading-system.com
    url: https://docs.trading-system.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.trading-system.com/v1
    description: Production server
  - url: https://staging-api.trading-system.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /strategies/{strategy_id}/backtest:
    post:
      tags: [Strategies]
      summary: Run backtest
      description: Execute a backtest for a specific strategy
      operationId: runBacktest
      parameters:
        - name: strategy_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BacktestRequest'
      responses:
        '202':
          description: Backtest started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestJob'
  /backtests/{job_id}:
    get:
      tags: [Strategies]
      summary: Get backtest results
      description: Retrieve results of a backtest job
      operationId: getBacktestResults
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Backtest results retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResults'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    PerformanceMetrics:
      type: object
      properties:
        total_return:
          type: number
        annualized_return:
          type: number
        volatility:
          type: number
        sharpe_ratio:
          type: number
        max_drawdown:
          type: number
        win_rate:
          type: number
        profit_factor:
          type: number

    # Risk Management Schemas
...
    BacktestRequest:
      type: object
      required: [start_date, end_date]
      properties:
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        initial_capital:
          type: number
          default: 100000
        benchmark:
          type: string
          default: SPY
    BacktestJob:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed]
        created_at:
          type: string
          format: date-time
        estimated_completion:
          type: string
          format: date-time
    BacktestResults:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [completed, failed]
        performance:
          $ref: '#/components/schemas/PerformanceMetrics'
        trades:
          type: array
          items:
            $ref: '#/components/schemas/Trade'
        equity_curve:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              value:
                type: number
    Trade:
      type: object
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
        price:
          type: number
        timestamp:
          type: string
          format: date-time
        pnl:
          type: number

    # Error Schema
...